Added test asserting Decrypt does not mutate
* Added a failing test asserting that Decrypt called on a NewDES3Cipher
  does not mutate the ciphertext.

The current des implementation reverses the order of dst & src from what
crypto (and idiomatic Go) expect. The test added in this commit expects
dst to come first, so it currently fails when Decrypt reads a zeroed
block from result, deciphers it, and stores it in enc. To make the test
pass, we'll need to switch the order of dst & src in the des
implementation.